py/requirements.txt
panda
numpy
paddlepaddle
loguru
tqdm
click